refactor(settings): extract SetsPage and SetEditorDialog
SetsPage is self-saving like the plan describes, but not via
profile_manager.add_set/delete_set (those exist but this UI path never
called them) -- it mutates the live set_decoders dict handed to it at
construction, same object identity as config_data["set_decoders"], so
edits are already in config_data by the time save_settings() runs.
